+++ WorldGenerator/NoiseGen/noiseGen.h
@@ -8,7 +8,7 @@ struct NoiseParameters
     int smoothness;
     int heightOffset;
 
-    double roughness;
+    float roughness;
 };
 
 class NoiseGenerator
@@ -17,18 +17,18 @@ class NoiseGenerator
     public:
         NoiseGenerator(int seed);
 
-        double getHeight(int x, int z, int chunkX, int chunkZ) const noexcept;
-        double getHeight(int x, int z) const noexcept;
+        float getHeight(int x, int z, int chunkX, int chunkZ) const noexcept;
+        float getHeight(int x, int z) const noexcept;
 
         void setParameters(const NoiseParameters& params) noexcept;
 
     private:
-        double getNoise(int  n) const noexcept;
-        double getNoise(double  x, double  z) const noexcept;
+        float getNoise(int  n) const noexcept;
+        float getNoise(float  x, float  z) const noexcept;
 
-        double lerp(double a, double b, double z) const noexcept;
+        float lerp(float a, float b, float z) const noexcept;
 
-        double noise(double  x, double  z) const noexcept;
+        float noise(float  x, float  z) const noexcept;
 
         NoiseParameters m_noiseParameters;
